start with navigator
