+++ src/DayPicker/Calendar/DaysCalendar/DaysCalendar.tsx
@@ -1,11 +1,8 @@
 import React from 'react';
 import styled from '@emotion/styled';
-import {css} from '@emotion/core';
 import {useDayPicker} from '../../DayPickerProvider';
-import {Span} from '../../../Typography';
-import {useTheme} from '../../../Themes';
-import {zeroPad} from '../utils/common';
-import {TCalendarDate} from '../../types';
+import {getDayTooltip} from '../utils/calendar';
+import {Day} from './Day';
 import {createDayCalendar} from './utils';
 
 const Days = styled.div`
@@ -16,81 +13,18 @@ const Days = styled.div`
   width: 100%;
   align-content: center;
   justify-items: center;
-  padding-top: 10px;
+  padding: 10px 16px 0px;
+
 `;
-
-const baseStyles = css`
-  display: flex;
-  width: 100%;
-  height: 100%;
-  justify-content: center;
-  align-items: center;
-  font-size: 12px;
-  cursor: pointer;
-  border-radius: 50%;
-`;
-
-type Props = {date: TCalendarDate};
-
-function Day({date}: Props) {
-  const {
-    handleChangeDay,
-    delimiter,
-    format,
-    day: currentDay,
-    month: currentMonth,
-    year: currentYear,
-  } = useDayPicker();
-  const theme = useTheme();
-  const {day, month, year, isCurrentMonth} = date;
-
-  const isCurrentDay = React.useCallback(() => {
-    return day === currentDay && month === currentMonth && year === currentYear;
-  }, [currentDay, currentMonth, currentYear, day, month, year]);
-
-  const dayStyles = css`
-    opacity: ${isCurrentMonth ? 1 : 0.5};
-    background-color: ${isCurrentDay() ? theme.colors.mainColors.primaryDark : 'transparent'};
-    color: ${isCurrentDay() ? theme.colors.textColors.textOnPrimary : theme.colors.textColors.text};
-    &:hover {
-      background-color: ${theme.colors.mainColors.primary};
-      color: ${theme.colors.textColors.textOnPrimary};
-    }
-  `;
-
-  const handleSelectDay = React.useCallback(() => {
-    const outputDate = [];
-    format.split(delimiter).forEach(ch => {
-      if (ch === 'DD') {
-        outputDate.push(zeroPad(day, 2));
-      } else if (ch === 'MM') {
-        outputDate.push(zeroPad(month, 2));
-      } else if (ch === 'YYYY') {
-        outputDate.push(year);
-      }
-    });
-    const newDate = outputDate.join(delimiter);
-    handleChangeDay(newDate);
-  }, [day, delimiter, format, handleChangeDay, month, year]);
-  return (
-    <Span
-      onClick={handleSelectDay}
-      css={[baseStyles, css({color: date.type === 'weekend' && '#D40000'}), dayStyles]}
-    >
-      {date.day}
-    </Span>
-  );
-}
 
 export function DaysCalendar() {
-  const {month, year} = useDayPicker();
-
+  const {month, year, locale} = useDayPicker();
   const calendar = createDayCalendar(month, year);
 
   return (
     <Days>
       {calendar.map(day => {
-        return <Day date={day} key={day.key} />;
+        return <Day date={day} key={day.key} title={getDayTooltip(day?.date, locale)} />;
       })}
     </Days>
   );
