+++ src/DayPicker/Calendar/Navigator/index.ts
@@ -1,0 +1,1 @@
+export * from './Navigator';
